🚚 move : 어댑터 파일 이동
